fix: queue API key propagation and fix unbound this in CES client
